requirements.txt
python-telegram-bot>=20.0
qrcode[pil]>=7.4.2
Pillow>=9.0.0
cryptography==41.0.7
requests==2.31.0
python-dotenv>=0.19.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
pyrogram>=2.0.0
TgCrypto>=1.2.0
Flask>=2.0
mysql-connector-python>=8.0